test(utils): revert page-status and site-admin unit tests
Remove test/page-status-utils.test.js and test/site-admin-utils.test.js.
Revert filterPendingPages back inline in diff.js and remove the
tools/page-status/utils.js extraction that existed only to support the
tests. getDAEditorURL bug fix in site-admin/helpers/utils.js is preserved.

Testing strategy for these utilities will be addressed in a separate PR.
